Add ispconfig to apache2
Fix: #574 
Since the poss_path is split on the last ```/``` character and find is recursive by default we shouldnt need to place a wildcard in the log file path. I have tested this and seems to be working, however, if the original issue owners are around they can test it on a system that would be great.