update fixes to be more consistent with the rest of the code
Upon review, the added code was pretty different in the approach
used to keep retrying compared to the approach for the QueryRange
method. I updated the method to create a new doTail that has the
same style as doQueryRange and updated Tail to use it. This has
the following effects:

- doTail will keep trying after losing a connection
- the retry interval will grow (with an upper limit) and shrink
  (with a lower limit) as connections are made and broken
- the time in the request is updated to avoid overlapping with
  previous data that was returned (missing in the first fix)