pkg/acquisition/modules/victorialogs: rework handling of backoff and "since" paramater
Simplify code of backoff handling in order to reduce duplication. Also do not sleep when making the first request in order to avoid artificial delay for startup.

While at it, implement proper handling of "since" parameter. Previously, it was reset to 0 and ignored in "tail" mode since VL API did not support tailing results from the past.
Implemented full support to use "since" value when performing initial tailing and keeping track of last seen log item in order to not miss log lines when retrying the request.